+++ backend/src/main/java/pl/mrs/movie_raiting_system/entities/User.java
@@ -40,7 +40,7 @@ public class User {
 
     @Transient
     private String passwordConfirm;
-    
+
     @Email
     @NotBlank
     @Column(name = "email")
