+++ backend/src/main/java/pl/mrs/movie_raiting_system/service/UserDetailsService.java
@@ -1,0 +1,4 @@
+package pl.mrs.movie_raiting_system.service;
+
+public class UserDetailsService {
+}
