+++ backend/src/main/java/pl/mrs/movie_raiting_system/service/UserServiceImpl.java
@@ -1,0 +1,26 @@
+package pl.mrs.movie_raiting_system.service;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import pl.mrs.movie_raiting_system.dao.UserRepository;
+import pl.mrs.movie_raiting_system.entities.User;
+
+public class UserServiceImpl implements UserService {
+
+    @Autowired
+    private UserRepository userRepository;
+
+    @Autowired
+    private BCryptPasswordEncoder bCryptPasswordEncoder;
+
+    @Override
+    public void save(User user) {
+        user.setPassword(bCryptPasswordEncoder.encode(user.getPassword()));
+        userRepository.save(user);
+    }
+
+    @Override
+    public User findByUsername(String username) {
+        return userRepository.findByUsername(username);
+    }
+}
